requirements.txt
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
importlib-metadata==1.7.0
iniconfig==1.0.1
jieba3k==0.35.1
joblib==0.16.0
lxml==4.5.2
more-itertools==8.4.0
newspaper3k==0.2.8
nltk==3.4.5
oauthlib==3.1.0
olefile==0.46
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.5
pytest==6.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
singledispatch==3.4.0.3
djangorestframework~=3.11.1